requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile '.\requirements.in'
#
asgiref==3.4.1
    # via django
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.7
    # via requests
cryptography==35.0.0
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
django==3.2.5
    # via
    #   -r .\requirements.in
    #   django-allauth
    #   django-contrib-comments
    #   django-filter
    #   django-mptt
    #   django-payments
    #   django-rest-auth
    #   djangorestframework
    #   pinax-badges
django-allauth==0.45.0
    # via -r .\requirements.in
django-categories==1.8
    # via -r .\requirements.in
django-contrib-comments==2.1.0
    # via
    #   -r .\requirements.in
    #   django-fluent-comments
    #   django-threadedcomments
django-countries==7.2.1
    # via -r .\requirements.in
django-crispy-forms==1.13.0
    # via django-fluent-comments
django-filter==21.1
    # via -r .\requirements.in
django-fluent-comments==3.0
    # via -r .\requirements.in
django-gamification==0.3.2
    # via -r .\requirements.in
django-js-asset==1.2.2
    # via django-mptt
django-mptt==0.11.0
    # via django-categories
django-payments[stripe]==0.15.0
    # via -r .\requirements.in
django-rest-auth==0.9.5
    # via -r .\requirements.in
django-rest-framework==0.1.0
    # via -r .\requirements.in
django-tag-parser==3.2
    # via django-fluent-comments
django-threadedcomments==1.2
    # via -r .\requirements.in
django-vote==2.2.0
    # via -r .\requirements.in
djangorestframework==3.12.4
    # via
    #   django-rest-auth
    #   django-rest-framework
idna==3.3
    # via requests
oauthlib==3.1.1
    # via requests-oauthlib
pillow==8.4.0
    # via -r .\requirements.in
pinax-badges==3.0.0
    # via -r .\requirements.in
psycopg2==2.8.4
    # via -r .\requirements.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.3.0
    # via django-allauth
python-akismet==0.4.2
    # via django-fluent-comments
python3-openid==3.2.0
    # via django-allauth
pytz==2021.3
    # via django
requests==2.26.0
    # via
    #   django-allauth
    #   django-payments
    #   python-akismet
    #   requests-oauthlib
    #   stripe
requests-oauthlib==1.3.0
    # via django-allauth
six==1.16.0
    # via
    #   django-rest-auth
    #   unicode-slugify
sqlparse==0.4.2
    # via django
stripe==2.62.0
    # via django-payments
symposion==1.0b2.dev2
    # via -r .\requirements.in
unicode-slugify==0.1.3
    # via django-categories
unidecode==1.3.2
    # via unicode-slugify
urllib3==1.26.7
    # via requests
